+++ app/javascript/recording_studio_notifications_push/controllers/push_devices_controller.js
@@ -7,7 +7,6 @@ export default class extends Controller {
   static values = {
     registerUrl: String,
     unregisterUrlTemplate: String,
-    testPushUrlTemplate: String,
     vapidKey: String,
     firebaseConfig: Object,
     firebaseReady: Boolean,